Add files via upload
> assignment2<-c(16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22)
>myMean<-function(assignment2) {return(sum(assignment)/length(someData))}

Run as it is initially written, the function 'myMean' contains an error with the relational operator and results in the following message: 
myMean <- function(assignment2) {return(sum(assignment)/length(someData))}
Error: unexpected '>' in ">"

The original code is assigning the function 'assignment2' to 'myMean' instead of assigning the equation for how to find the mean of assignment2 to 'myMean'.  By writing this equation as an expression instead of the function, the code is stating a generalization of finding mean but is not associating the generalization with the elements of assignment2. The verbiage of the code is also wrong as it reads the "sum of 'assignment' divided by the length of 'someData'" with neither of these arguments having been assigned.

There are several ways to find the mean of 'assignment2' in R, however, the code closest to the original code's format would read
>assignment2<-c(16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22)
>myMean <-  sum(assignment2)/length(assignment2)